refactor: 선착순 로직 인터페이스-구체클래스 분리
- DI를 통한 자유로운 구현체 사용
